Added EditorJS

+++ src/components/EditPost.jsx
@@ -1,10 +1,13 @@
 import React from "react";
 import { Redirect } from "react-router-dom";
 import Loader from "./Loader";
+import Editor from "./Editor";
 import NotFound from "./NotFound";
+
 import getRequest from "../utils/getRequest";
 import putRequest from "../utils/putRequest";
 import deleteRequest from "../utils/deleteRequest";
+import editorConfig from "../utils/editorConfig";
 
 class EditPost extends React.Component {
 	constructor(props) {
@@ -24,6 +27,7 @@ class EditPost extends React.Component {
 				body: "",
 			},
 		};
+		this.editor = null;
 	}
 
 	changeHandler = ({ target }) => {
@@ -62,18 +66,22 @@ class EditPost extends React.Component {
 
 	deleteHandler = async (e) => {
 		const { slug } = this.props;
-		const { article, errors } = await deleteRequest(`/api/articles/${slug}`);
-		this.setState({
-			isDeleted: article ? true : false,
-			error: errors ? true : false
-		});
-	}
+		const confirmation = window.confirm("Do You Want to Delete this Article");;
+		if (confirmation) {
+			const { article, errors } = await deleteRequest(`/api/articles/${slug}`);
+			this.setState({
+				isDeleted: article ? true : false,
+				error: errors ? true : false,
+			});
+		}
+		return;
+	};
 
   async componentDidMount () {
-    const { slug } = this.props;
+		const { slug } = this.props;
 		const { article, errors } = await getRequest(`/api/articles/${slug}`);
-		console.log(errors, "mounting of edit page");
-    this.setState({
+		console.log("mounting of edit page");
+		this.setState({
 			title: article?.title ?? "",
 			description: article?.description ?? "",
 			body: article?.body ?? "",
@@ -81,32 +89,44 @@ class EditPost extends React.Component {
 			author: article?.author.username,
 			error: errors ? true : false,
 		});
-  }
+		const blocks = JSON.parse(article.body).blocks
+		this.editor = editorConfig("body", blocks);
+		console.log(this.editor, JSON.parse(article.body))
+	}
+
+	async componentDidUpdate () {
+	}
+
 
   render () {
-		const { isLoggedIn, user } = this.props
+		const { user } = this.props;
     const { title, description, body, tagList, error, author, article, isDeleted } = this.state;
 
 		if (isDeleted) return <Redirect to="/" />;
 		if (error) return <NotFound />;
-		if (!(title && description && body) || !(isLoggedIn && user)) return <Loader />;
+		if (!(title && description && body)) return <Loader />;
 		if (user.username !== author) return <Redirect to="/" />;
 		if (article) return <Redirect to={`/article/${article.slug}`} />;
 
 		return (
 			<div className="container mx-auto p-4">
 				<form className="w-3/4 mx-auto p-6 rounded-2xl bg-gray-50 shadow-lg" onSubmit={this.editHandler}>
-					<h2 className="text-2xl uppercase text-gray-500 text-center m-8">Update Article</h2>
+					<h2 className="text-2xl uppercase text-gray-500 text-center m-8">Create Article</h2>
 					<label htmlFor="title">Title -</label>
-					<input type="text" name="title" id="title" minLength="15" placeholder="Title should be minimum 15 character" value={title} onChange={this.changeHandler} required />
+					<input type="text" name="title" id="title" minLength="15" placeholder="Title should be minimum 15 character" value={this.state.title} onChange={this.changeHandler} required />
+					<span className="text-red-700 text-sm">{this.state.errors.title || ""}</span>
 					<label htmlFor="description">Description -</label>
-					<input type="text" name="description" id="description" minLength="25" placeholder="Description should be minimum 20 character" value={description} onChange={this.changeHandler} required />
-					<label htmlFor="body">Content -</label>
-					<textarea name="body" id="body" cols="30" rows="10" minLength="100" placeholder="Content should be minimum of 100 character" value={body} onChange={this.changeHandler} required />
+					<input type="text" name="description" id="description" minLength="25" placeholder="Description should be minimum 20 character" value={this.state.description} onChange={this.changeHandler} required />
+					<span className="text-red-700 text-sm">{this.state.errors.description || ""}</span>
+					<label htmlFor="description">Content -</label>
+					<Editor />
+					<span className="text-red-700 text-sm">{this.state.errors.body || ""}</span>
 					<label htmlFor="tagList">Tags -</label>
-					<input type="text" name="tagList" id="tagList" placeholder="Multiple tags should be separated by commas (cars,books)" value={tagList} onChange={this.changeHandler} />
+					<input type="text" name="tagList" id="tagList" placeholder="Multiple tags should be separated by commas (cars,books)" value={this.state.tagList} onChange={this.changeHandler} />
 					<div className="text-right">
-						<button className="btn del-btn mx-4" onClick={this.deleteHandler}>Delete</button>
+						<button className="btn del-btn mx-4" onClick={this.deleteHandler}>
+							Delete
+						</button>
 						<input type="submit" value="Update Post" className="btn" />
 					</div>
 				</form>
